+++ spec/rspec/core/example_group_spec.rb
@@ -441,7 +441,7 @@ module RSpec::Core
         context "and a string in an inner group" do
           it "is the top level constant" do
             group = ExampleGroup.describe(String) do
-              describe :symbol do
+              describe "inner" do
                 example "described_class is String" do
                   expect(described_class).to eq(String)
                 end
@@ -457,7 +457,7 @@ module RSpec::Core
             group = ExampleGroup.describe(String) do
               described_class
               metadata[:described_class] = Object
-              describe :symbol do
+              describe "inner" do
                 example "described_class is Object" do
                   expect(described_class).to eq(Object)
                 end
@@ -1027,7 +1027,7 @@ module RSpec::Core
         end
 
         it 'generates a pending example group' do
-          group = ExampleGroup.send(method_name) do
+          group = ExampleGroup.send(method_name, "group") do
             it("passes") { }
             it("fails")  { expect(2).to eq(3) }
           end
@@ -1081,7 +1081,7 @@ module RSpec::Core
       end
 
       it 'marks every example as pending' do
-        group = ExampleGroup.describe(:pending => true) do
+        group = ExampleGroup.describe("group", :pending => true) do
           it("passes") { }
           it("fails", :pending => 'unimplemented')  { fail }
         end